Logging the game (logs.txt) | opening beta-testing
Welcome to beta-testing, guys!

Now it is your time. You can download wwtbam_game.exe and test the game now! You need to push and click everywhere, test aggressively and by your own user experience. If you catch an error, you will get a message saying that you should send logs.txt to me in Telegram, but during this time of open testing, it is better to create an issue in GitHub (github.com/PaveTranquil/WWBM/issues/new), if you are registered there, and describe the problem with the steps that led to its reproduction. Else, of course, write to me in Telegram (t.me/pavetranquil)